+++ anybox/buildbot/openerp/subfactories/postdownload.py
@@ -1,0 +1,2 @@
+def noop(configurator, options, buildout_slave_path, environ=()):
+    return buildout_slave_path, ()